Update badge URLs
I went to copy the badge link over but somehow they 404 for me. I replaced them with the raw link generated by clicking the "raw" button on the file on github. I also added the full link to the badge so you can copy the full markup over to a README