Update clang-format with InsertBraces: true
A very common review request is to add braces to if statements even if
they only have a single statement. Not doing so is a subtle source of
bugs, hece we are upating the clang-format config to automatically add
missing braces in such situations.

See clang-format doc for more details:
https://releases.llvm.org/19.1.0/tools/clang/docs/ClangFormatStyleOptions.html#insertbraces